requirements.txt
captum==0.8.0
ConfigArgParse==1.7
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.5
pandas==2.2.3
pytorch_lightning==2.5.1.post0
PyYAML==6.0.2
scikit_learn==1.6.1
torch==2.7.0
torchmetrics==1.7.1
tqdm==4.64.1
wandb==0.19.10
plotly==5.10.0
seaborn==0.12.1